asset : 영웅 스킬 구현
Shieldman 기절 스킬 구현 및 fx 적용
장비 id int로 변경

+++ Assets/Scripts/Hero/Equipment/EquipmentManager.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public class EquipmentManager
 {
-    public static Dictionary<long, Equipment> EquipDic = new Dictionary<long, Equipment>();
+    public static Dictionary<int, Equipment> EquipDic = new Dictionary<int, Equipment>();
 }
